+++ packages/map-template/src/components/MapWrapper/MapWrapper.jsx
@@ -22,6 +22,7 @@ import useMapBoundsDeterminer from '../../hooks/useMapBoundsDeterminer';
 import hideNonMatchesState from '../../atoms/hideNonMatchesState';
 import PropTypes from 'prop-types';
 import ViewSelector from '../ViewSelector/ViewSelector';
+import LanguageSelector from '../LanguageSelector/LanguageSelector.jsx';
 import appConfigState from '../../atoms/appConfigState';
 import isNullOrUndefined from '../../helpers/isNullOrUndefined';
 
@@ -34,7 +35,9 @@ MapWrapper.propTypes = {
     resetCount: PropTypes.number.isRequired,
     mapOptions: PropTypes.object,
     gmMapId: PropTypes.string,
-    isWayfindingOrDirections: PropTypes.bool
+    isWayfindingOrDirections: PropTypes.bool,
+    currentLanguage: PropTypes.string,
+    setLanguage: PropTypes.func
 };
 
 /**
@@ -57,9 +60,11 @@ let _tileStyle;
  * @param {object} props.mapOptions - Options for instantiating and styling the map as well as UI elements.
  * @param {string} props.gmMapId - Google Maps Map ID for custom styling.
  * @param {boolean} props.isWayfindingOrDirections - Whether wayfinding or directions is active or not.
+ * @param {string} props.currentLanguage - The currently selected language code.
+ * @param {function} props.setLanguage - Function to set the selected language.
  * @returns
  */
-function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule, onMapPositionInvestigating, onViewModeSwitchKnown, resetCount, mapOptions, gmMapId, isWayfindingOrDirections }) {
+function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule, onMapPositionInvestigating, onViewModeSwitchKnown, resetCount, mapOptions, gmMapId, isWayfindingOrDirections, currentLanguage, setLanguage }) {
     const apiKey = useRecoilValue(apiKeyState);
     const gmApiKey = useRecoilValue(gmApiKeyState);
     const mapboxAccessToken = useRecoilValue(mapboxAccessTokenState);
@@ -78,6 +83,7 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
     const hideNonMatches = useRecoilValue(hideNonMatchesState);
     const appConfig = useRecoilValue(appConfigState);
     const [isViewSelectorVisible, setIsViewSelectorVisible] = useState(false);
+    const [isLanguageSelectorVisible, setIsLanguageSelectorVisible] = useState(false);
 
     useLiveData(apiKey);
 
@@ -296,7 +302,7 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
     }, [tileStyle]);
 
     /**
-     * React on changes in appConfig and sets visibility of View Selector.
+     * React on changes in appConfig and sets visibility of View Selector and visibility of Language Selector.
      */
     useEffect(() => {
         if (appConfig) {
@@ -305,6 +311,14 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
             } else {
                 // Boolean from the App Config comes as a string. We need to return clean boolean value based on that.
                 setIsViewSelectorVisible(appConfig?.appSettings?.viewSelector.trim().toLowerCase() === 'true');
+            }
+
+
+            if (isNullOrUndefined(appConfig?.appSettings?.languageSelector)) {
+                setIsLanguageSelectorVisible(false);
+            } else {
+                // Boolean from the App Config comes as a string. We need to return clean boolean value based on that.
+                setIsLanguageSelectorVisible(appConfig?.appSettings?.languageSelector.trim().toLowerCase() === 'true');
             }
         }
     }, [appConfig])
@@ -320,7 +334,10 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
             gmMapId={gmMapId}
         />}
         {/* Pass isWayfindingOrDirections prop to ViewSelector to disable interactions while wayfinding or directions is active*/}
-        {apiKey && < ViewSelector isViewSelectorDisabled={isWayfindingOrDirections} isViewSelectorVisible={isViewSelectorVisible} />}
+        {apiKey && <>
+            <ViewSelector isViewSelectorVisible={isViewSelectorVisible} isViewSelectorDisabled={isWayfindingOrDirections} />
+            <LanguageSelector currentLanguage={currentLanguage} setLanguage={setLanguage} isVisible={isLanguageSelectorVisible} />
+        </>}
     </>)
 }