+++ packages/map-template/src/i18n/zh.js
@@ -79,7 +79,9 @@ const zh = {
     'Open': '营业中',
     'Closed': '已关闭',
     // View Selector
-    'Go to view': '定位到'
+    'Go to view': '定位到',
+    // Language Selector
+    'Select language': '选择语言',
 };
 
 export default zh;